requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.1
Automat==25.4.16
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
Django==5.2.7
django-encrypted-model-fields==0.6.5
django-restframework==0.0.1
django-unfold==0.68.0
djangorestframework==3.16.1
dotenv==0.9.9
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jiter==0.11.1
msgpack==1.1.2
oauthlib==3.3.1
openai==2.5.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.3
redis==7.0.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
zope.interface==8.0.1
